feat!: Support custom handles on the new PDS (#49)
Closes #26 

Add support for using a custom handle on the new PDS. This is
accomplished by calling the `com.atproto.identity.updateHandle` endpoint
as the final step of the migration. Because a temporary non-custom
handle has to be used to create the account in the first place, required
breaking the format of the migration credentials.